web(components): modulariza o header das páginas
