+++ web/app/schedules/layout.tsx
@@ -1,71 +1,40 @@
 'use client';
 
-import { useState } from 'react';
 import { usePathname, useRouter } from 'next/navigation';
 import useUser from '@/app/hooks/useUser';
 
 import AsideButton from '../components/AsideButton';
 import Protected from '../components/Protected';
 import { LoadingScreen } from '../components/LoadingScreen';
+import InfoHeader from '../components/InfoHeader';
 
 import homeIcon from '@/public/icons/home.jpg';
 import scheduleIcon from '@/public/icons/schedule.jpg';
 import profileIcon from '@/public/icons/profile.jpg';
-
-
-const days = ['Domingo', 'Segunda', 'Terça', 'Quarta', 'Quinta', 'Sexta', 'Sábado'];
-const months = ['Janeiro', 'Fevereiro', 'Março', 'Abril', 'Maio', 'Junho', 'Julho', 'Agosto', 'Setembro', 'Outubro', 'Novembro', 'Dezembro'];
-
-const pathPX = {
-    'home': 'left-[18px]',
-    'mygrades': 'left-[99px]',
-    'profile': 'left-[175px]',
-};
-
-const pathPXAnonymous = {
-    'home': 'left-[18px]',
-    'mygrades': 'left-[132px]'
-};
-
-const currentDateObject = new Date(
-    new Date().toLocaleString('en', {
-        timeZone: 'America/Sao_Paulo'
-    })
-);
+import useWindowDimensions from '../hooks/useWindowDimensions';
 
 function LayoutJSX({ children }: { children: React.ReactNode }) {
+    const { breakHeighPoint } = useWindowDimensions();
     const { user, isLoading } = useUser();
-    const router = useRouter();
-    const [currentDate, _] = useState(currentDateObject);
-    const path = usePathname().split('/')[2] as keyof typeof pathPX;
 
-    const currentDay = currentDate.getDate().toString();
-    const currentMonth = months[currentDate.getMonth()];
-    const currentWeekDay = days[currentDate.getDay()];
+    const router = useRouter();
+    const path = usePathname().split('/')[2];
 
     if (isLoading) return <LoadingScreen />;
 
     return (
         <>
-            <header className='rounded-b-[40px] mb-8 flex flex-col justify-end bg-primary px-6 h-24'>
-                <h1 className='col-span-2 font-semibold text-base text-white'>
-                    Olá, {user.is_anonymous ? 'Anônimo' : user.first_name}!
-                </h1>
-                <p className='mb-5 col-span-2 font-medium text-base text-white'>
-                    {currentWeekDay}, {currentDay} de {currentMonth}
-                </p>
-            </header>
-
-            <main className='pt-3 h-[calc(100%-17.75rem)]'>
+            <InfoHeader />
+            <main className={`${breakHeighPoint ? 'pt-3 h-[calc(100%-15.75rem)]' : 'pt-7 h-[calc(100%-9.75rem)]'}`}>
                 {children}
             </main>
-
-            <div className="justify-around flex bg-white rounded-full px-6 py-2 w-[275px] absolute m-auto inset-x-px bottom-8 backdrop-blur-sm bg-opacity-50 drop-shadow-lg">
-                <div
-                    className={`bg-primary transition-all duration-300
-                    ${user.is_anonymous ? 'w-[120px]' : 'w-[85px]'} h-[50px]
-                    rounded-full absolute ${user.is_anonymous ? pathPXAnonymous[path as keyof typeof pathPXAnonymous] : pathPX[path]}`}
-                ></div>
+            <div className="flex justify-around bg-blue-500 rounded-t-[40px] px-6 py-2 max-w-xl absolute m-auto inset-x-px bottom-0 backdrop-blur-sm bg-opacity-50 drop-shadow-lg">
+                <section className='grid grid-cols-3 px-6 absolute w-full'>
+                    <div className={
+                        `h-[50px] rounded-full bg-primary transition-all duration-300
+                        ${path === 'home' ? 'col-start-1' : (path === 'mygrades' ? 'col-start-2' : 'col-start-3')}
+                    `}></div>
+                </section>
                 <AsideButton image={homeIcon} text='Home' onClick={() => router.push('/schedules/home')} />
                 <AsideButton image={scheduleIcon} text='Grades' onClick={() => router.push('/schedules/mygrades')} />
                 {user.is_anonymous ? null : <AsideButton image={profileIcon} text='Perfil' onClick={() => router.push('/schedules/profile')} />}
